initital test and clean
It runs. Now gotta check if its valid
