+++ trgt/database/__init__.py
@@ -1,12 +1,10 @@
 """ database commands """
-from trgt.database.create import (
-    create_main,
-    vcf_to_tdb
-)
-
+from trgt.database.create import create_main
 from trgt.database.query import query_main
+from trgt.database.append import append_main
 
 __all__ = [
     "create_main",
-    "query_main"
+    "query_main",
+    "append_main"
 ]
